merge

+++ src/main/java/im/cave/ms/client/character/items/Inventory.java
@@ -8,21 +8,11 @@ import im.cave.ms.tools.Util;
 import lombok.Getter;
 import lombok.Setter;
 
-import javax.persistence.CascadeType;
-import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.OneToMany;
-import javax.persistence.Table;
+import javax.persistence.*;
 import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.List;
 import java.util.concurrent.CopyOnWriteArrayList;
-import java.util.stream.Collectors;
 
 /**
  * @author fair
@@ -61,7 +51,8 @@ public class Inventory {
         return items;
     }
 
-    public Item getItem(short pos) {
+
+    public Item getItem(int pos) {
         return getItems().stream().filter(item -> item.pos == pos).findAny().orElse(null);
     }
